requirements.txt
amqp==5.3.1
celery==5.4.0
Django==5.1.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
stripe==4.0.2

